// frontend/src/api/generated/models/picking-instruction-detail.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { PickingInstructionLineDetail } from './picking-instruction-line-detail';
// May contain unused imports in some cases
// @ts-ignore
import type { PickingInstructionStatus } from './picking-instruction-status';

/**
 * 
 * @export
 * @interface PickingInstructionDetail
 */
export interface PickingInstructionDetail {
    /**
     * ピッキング指示ID
     * @type {number}
     * @memberof PickingInstructionDetail
     */
    'id'?: number;
    /**
     * 指示番号
     * @type {string}
     * @memberof PickingInstructionDetail
     */
    'instructionNumber'?: string;
    /**
     * 倉庫ID
     * @type {number}
     * @memberof PickingInstructionDetail
     */
    'warehouseId'?: number;
    /**
     * 倉庫名
     * @type {string}
     * @memberof PickingInstructionDetail
     */
    'warehouseName'?: string;
    /**
     * エリアID
     * @type {number}
     * @memberof PickingInstructionDetail
     */
    'areaId'?: number | null;
    /**
     * エリア名
     * @type {string}
     * @memberof PickingInstructionDetail
     */
    'areaName'?: string | null;
    /**
     * 
     * @type {PickingInstructionStatus}
     * @memberof PickingInstructionDetail
     */
    'status'?: PickingInstructionStatus;
    /**
     * 
     * @type {Array<PickingInstructionLineDetail>}
     * @memberof PickingInstructionDetail
     */
    'lines'?: Array<PickingInstructionLineDetail>;
    /**
     * 作成日時
     * @type {string}
     * @memberof PickingInstructionDetail
     */
    'createdAt'?: string;
    /**
     * 作成者ID
     * @type {number}
     * @memberof PickingInstructionDetail
     */
    'createdBy'?: number;
    /**
     * 作成者名
     * @type {string}
     * @memberof PickingInstructionDetail
     */
    'createdByName'?: string;
    /**
     * 完了日時
     * @type {string}
     * @memberof PickingInstructionDetail
     */
    'completedAt'?: string | null;
    /**
     * 完了者ID
     * @type {number}
     * @memberof PickingInstructionDetail
     */
    'completedBy'?: number | null;
}