// frontend/src/api/generated/models/picking-instruction-status.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * ピッキング指示ステータス - CREATED: 作成済み - IN_PROGRESS: 作業中 - COMPLETED: 完了 
 * @export
 * @enum {string}
 */

export const PickingInstructionStatus = {
    Created: 'CREATED',
    InProgress: 'IN_PROGRESS',
    Completed: 'COMPLETED'
} as const;

export type PickingInstructionStatus = typeof PickingInstructionStatus[keyof typeof PickingInstructionStatus];